+++ WebFinalProject/Result.cs
@@ -1,0 +1,14 @@
+namespace WebFinalProject;
+
+public record Result<T>(
+    T Data,
+    bool IsSuccess,
+    string Reason
+)
+{
+    public static Result<T> Ok(T data)
+        => new(data, true, null);
+
+    public static Result<T> BadRequest(string reason)
+        => new(default, false, reason);
+}
