+++ WebFinalProject/UseCases/Login/LoginRequest.cs
@@ -1,11 +1,10 @@
 using System.ComponentModel.DataAnnotations;
 using WebFinalProject.Models;
+using WebFinalProject.Validations;
 
 namespace WebFinalProject.UseCases.Login;
 
-public record LoginRequest
-{
-    // validação pra ver se o nome de usuário já existe
-    [Required]
-    [MinLength()]
-}
+public record LoginRequest(
+    string Login,
+    string Password
+);
