+++ WebFinalProject/UseCases/Login/LoginResponse.cs
@@ -1,14 +1,5 @@
-namespace WebFinalProject;
+namespace WebFinalProject.UseCases.Login;
 
-public record Result<T>(
-    T Data,
-    bool IsSuccess,
-    string Reason
-)
-{
-    public static Result<T> Ok(T data)
-        => new(data, true, null);
-
-    public static Result<T> BadRequest(string reason)
-        => new(default, false, reason);
-}
+public record LoginResponse(
+    string Token
+);
